Allow tasks to be aborted
Right now, aborting a task is similar to pausing a task.
However, aborted tasks will not be resumable (unlike paused tasks).

This PR also updates the logic in task.rb to check the run status
prior to running an iteration, and ensures we only update status
on the run in `job_running` if the run is not already paused or
aborted.