Don't load rails/all in bin/rails
Loading rails/all so early prevents the test_unit railtie from setting
RAILS_ENV, because at that point we're not in a Rake application yet.

It's not an issue for `$ rails test` because it loads the TestCommand
which uses the TestUnit::Runner which sets RAILS_ENV to test.

https://github.com/rails/rails/blob/v6.0.3.4/railties/lib/rails/commands/test/test_command.rb#L32
https://github.com/rails/rails/blob/v6.0.3.4/railties/lib/rails/test_unit/runner.rb#L26

For `$ rails test:system` though, the Rails command infrastructure falls
back to considering this is a Rake task:

https://github.com/rails/rails/blob/v6.0.3.4/railties/lib/rails/command.rb#L48

which loads the Engine Rakefile, which loads the dummy application
Rakefile, which loads the test_unit railtie. At this point, the
RAILS_ENV will be set because there will be a Rake application and the
task starts with test:, like mentioned in the previous commit.

This only happens if the test_unit railtie hasn't been already loaded
though, which happens because of the rails/all.

The normal bin/rails command in a Rails application doesn't do this, and
it doesn't seem necessary for a Rails engine either.