Remove TestTask class leak
Defining a Task in a test class results in the Task being leaked across
the test suite. Since we list the descendants of Task in
Task.available_task, as soon as the test file defining the class is
loaded, other tests expect the Task to exist.

This is why we have TaskDataTest referring to the TestTask, meaning this
test file can't be run by itself, since it needs the other test file to
be loaded as well to define the TestTask.

When we run all the tests using the Rake task, it adds another
difficulty where the system tests now fail because the TestTask now is
in the list of task, when it's not when TasksTest runs individually or
with rails test:system.

This removes the TestTask defined in TaskJobTest to replace it with a
regular Task that is always present.