Remove EXISTS query on tasks#show
Since we're paginating, we're already loading the completed runs, so we can
avoid one query by checking there's either an active run or a completed run
before checking if it's a deleted task.

This is done by inlining TaskDataShow.find into the controller and then moving
the Task.named call at the very last moment.

This has a small implication, if you "paginate" past the last cursor of a
deleted task, you'll get a 404 instead of the empty task page with the last run
at the top. This is not accessible via links anyway so it's fine.

Co-authored-by: Adrianna Chang <adrianna.chang@shopify.com>